Add CategoryPageFolder and rename categories according to API
